+++ nautilus_core/model/src/python/events/position/mod.rs
@@ -13,8 +13,4 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-//! Events for the trading domain model.
-
-pub mod account;
-pub mod order;
-pub mod position;
+pub mod snapshot;